[upstream][paper][touchhandler] Fix mouseUp events skipping touch handler with event tracking loops
Summary:
AppKit NSTextField uses an event tracking run loop on mouse down, waiting for and dequeuing the mouse up event to enable editing of the field.

The dequeuing of the mouse up event means it won't be submitted to the touch handler and leaving an ongoing touch session started by the mouse down.

This diff catches the deque and sends it asynchronously to let the touch handler process it after the event tracking exited.

This fixes press events for views located above a text field. The change will submit the mouse down and up event, allowing the press event to be triggered.

Test Plan:
- Run Messenger Desktop
- Log out
- Log in with email and password
- Enter a password and click on the "show password" button

| Before | After |
|--|
|   https://pxl.cl/4G0gc   |  https://pxl.cl/4G0fV  |

- Log in
- Search and clear the search by clicking on the discard icon

 https://pxl.cl/4G0kC

Reviewers: lyahdav, shawndempsey, #rn-desktop

Reviewed By: lyahdav

Subscribers: ericroz, ngerlem

Differential Revision: https://phabricator.intern.facebook.com/D55951356

Tasks: T183747164, T184150391