[upstream][paper] Support using default image loader request handling from image URL loader
Summary:
This diff extends the `RCTImageURLLoader` API to allow the `RCTImageLoader` to handle the request handling while letting the matching `RCTImageURLLoader` make decisions on the caching and the URL to use for the request.

Before this change, the full request/image loading implementation had to be handled by the `RCTImageURLLoader`, which would generate a lot of duplicated code.

Test Plan: Tested in this stack: D56172333

Reviewers: lyahdav, bumbu, #rn-desktop

Reviewed By: lyahdav

Differential Revision: https://phabricator.intern.facebook.com/D56172172

Tasks: T184302336